extract CompactionState
