+++ core/src/main/java/org/apache/druid/timeline/CompactionState.java
@@ -1,0 +1,85 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package org.apache.druid.timeline;
+
+import com.fasterxml.jackson.annotation.JsonCreator;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import org.apache.druid.indexer.partitions.PartitionsSpec;
+
+import java.util.Map;
+import java.util.Objects;
+
+public class CompactionState
+{
+  private final PartitionsSpec partitionsSpec;
+  // org.apache.druid.segment.IndexSpec cannot be used here to avoid the dependency cycle
+  private final Map<String, Object> indexSpec;
+
+  @JsonCreator
+  public CompactionState(
+      @JsonProperty("partitionsSpec") PartitionsSpec partitionsSpec,
+      @JsonProperty("indexSpec") Map<String, Object> indexSpec
+  )
+  {
+    this.partitionsSpec = partitionsSpec;
+    this.indexSpec = indexSpec;
+  }
+
+  @JsonProperty
+  public PartitionsSpec getPartitionsSpec()
+  {
+    return partitionsSpec;
+  }
+
+  @JsonProperty
+  public Map<String, Object> getIndexSpec()
+  {
+    return indexSpec;
+  }
+
+  @Override
+  public boolean equals(Object o)
+  {
+    if (this == o) {
+      return true;
+    }
+    if (o == null || getClass() != o.getClass()) {
+      return false;
+    }
+    CompactionState that = (CompactionState) o;
+    return Objects.equals(partitionsSpec, that.partitionsSpec) &&
+           Objects.equals(indexSpec, that.indexSpec);
+  }
+
+  @Override
+  public int hashCode()
+  {
+    return Objects.hash(partitionsSpec, indexSpec);
+  }
+
+  @Override
+  public String toString()
+  {
+    return "CompactionState{" +
+           "partitionsSpec=" + partitionsSpec +
+           ", indexSpec=" + indexSpec +
+           '}';
+  }
+}
